refactor(s3-bucket): query on region vs querying region separate
